rename `Workspace.Workspace` to `Windows.Sidebar`
