+++ Ktisis/Interface/Workspace/Manipulable.cs
@@ -1,7 +1,7 @@
 ﻿using System.Numerics;
 using System.Collections.Generic;
 
-namespace Ktisis.Interface.Editor {
+namespace Ktisis.Interface.Workspace {
 	public abstract class Manipulable {
 		public static Vector4 RootObjectCol = new Vector4(102, 226, 110, 255) / 255f;
 		public static Vector4 SubCategoryCol = new Vector4(8, 128, 255, 255) / 255f;
